requirements.txt
crewai==0.28.0
langchain==0.1.0
langchain-openai==0.0.5
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
streamlit==1.28.1
pandas==2.1.4
numpy==1.25.2
requests==2.31.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 